add intelligent macro name generator function
It attempts to use the struct name with the first letter capitalized,
if this results in the same name, then it capitalizes the entire string.

Thus, it will be:

ether => Ether
Udp => UDP
Ip => IP
iPv6 => IPv6
Tcp => TCP
someFunkyProtocol => SomeFunkyProtocol

Given that the macros will be used much more than the struct name,
this kind of heuristic should be ok.